backend/requirements.txt
# GYAAN-AI Backend Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.2

# AI/ML
groq==0.4.2
openai==1.6.1
langchain==0.0.350
langchain-openai==0.0.2

# PostgreSQL Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23
psycopg2-binary==2.9.9

# Database (optional for production)
supabase==2.3.0

# Utilities
python-dotenv==1.0.0
httpx>=0.24.0,<0.25.0
